+++ src/seekstorm/tokenizer.rs
@@ -133,6 +133,7 @@ pub(crate) fn tokenizer(
                             }
                             true
                         }
+
                         _ => {
                             if start {
                                 non_unique_terms_line.push(&text_normalized[start_pos..char.0]);
@@ -142,6 +143,7 @@ pub(crate) fn tokenizer(
                     };
                 }
             }
+
             TokenizerType::UnicodeAlphanumeric => {
                 text_normalized = text.to_lowercase();
                 for char in text_normalized.char_indices() {
@@ -152,6 +154,7 @@ pub(crate) fn tokenizer(
                             }
                             true
                         }
+
                         '"' | '+' | '-' | '#' => {
                             if !start {
                                 start_pos = char.0;
@@ -183,6 +186,7 @@ pub(crate) fn tokenizer(
                             }
                             true
                         }
+
                         _ => {
                             let apostroph = APOSTROPH.contains(&char.1);
                             if start {