+++ CHANGELOG.md
@@ -5,11 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.12.16] - 2025-02-14
+
+### Added
+
+- Basic tests added (issue #33): cargo test
+- New method current_doc_count() returns the number of indexed documents - deleted documents.
+
+### Fixed
+
+- Fixed issue #32 in clear_index.
+
 ## [0.12.15] - 2025-02-12
 
 ### Fixed
 
-- Fixes issue #34 - refactoring of http_server (by @gabriel-v)
+- Fixed issue #34 - refactoring of http_server (by @gabriel-v)
 
 ## [0.12.14] - 2025-02-10